Implement the type checker for Categories.
